api/requirements.txt
anyio==3.6.1
async-timeout==4.0.2
beanie==1.11.6
click==8.1.3
Deprecated==1.2.13
fastapi==0.78.0
h11==0.13.0
idna==3.3
motor==3.0.0
multidict==6.0.2
packaging==21.3
pydantic==1.9.1
pymongo==4.1.1
pyparsing==3.0.9
python-dotenv==0.20.0
redis==4.3.4
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
tornado==6.1
typing_extensions==4.2.0
uvicorn==0.18.2
wrapt==1.14.1
yarl==1.7.2
